Add background image to homeview
Implement background image feature as per [1]. The patch adds
homebackgroundbox.py and introduces a new control panel widget for
selecting a background image.

Some details:

(1) Two gconf settings are used to tell homebackgroundbox.py which
image to use for the background and what alpha setting to use when
rendering the image.

(2) Currently the control panel widget only allows the user to set the
background image; a reasonable default alpha setting (75%) is used.

(3) The images available for the background are looked for recursively
in ~/Documents and /usr/share/backgrounds .

(4) The gconf setting is being monitored by the background box, so
updating the setting automatically updates the background.

Future enhancement: expose some mechanism for setting the alpha of the
background.

This series of patches is based on the work of Agustin Zubiaga.

[1] http://wiki.sugarlabs.org/go/Features/Background_image_on_home_view