Complete port to introspection; fix problem with language setting
This series of patches completes the conversion to intropection for
the keyboard section of the Sugar control panel. The issues were: (1)
conversion to XKL (import xklavier was causing Sugar to crash); (2)
work-around for the problem with gconf get_list and set_list (dnarvaez
added a Sugar extention to enable set_list to save a list of keyboard
layouts); and (3) fixing a problem with setting the proper default
languages in the UI.

Part 3 requires a conversion from ISO-639 2-character codes to ISO-639
3-character codes. This is accomplished by making use of a table
available at:
http://www.loc.gov/standards/iso639-2/ISO-639-2_utf-8.txt
A snapshot of this table is now included in sugar/data.